+++ perl/FSM/Support/CapabilityManifestContract.pm
@@ -229,6 +229,8 @@ sub capability_manifest_embedding_keys {
             composition_report
             hdl_generator_facade
             hdl_generator_result
+            isf_public_interface
+            serializable_plan_reports
             typed_extensions
             debug_runtime
             section_contract