+++ perl/FSM/Support/EmbeddingSection.pm
@@ -11,6 +11,7 @@ use FSM::Support::EmbeddingContract qw(build_embedding_contract);
 use FSM::Support::ExtensionContract qw(build_extension_contract);
 use FSM::Support::HDLGeneratorFacadeContract qw(build_hdl_generator_facade_contract);
 use FSM::Support::HDLGeneratorResultContract qw(build_hdl_generator_result_contract);
+use FSM::Support::ISFPublicInterfaceContract qw(build_isf_public_interface_contract);
 use FSM::Support::SerializablePlanReportContract qw(build_serializable_plan_report_contract);
 
 our @EXPORT_OK = qw(
@@ -22,6 +23,7 @@ sub build_embedding_section {
         composition_report => build_composition_report_contract(),
         hdl_generator_facade => build_hdl_generator_facade_contract(),
         hdl_generator_result => build_hdl_generator_result_contract(),
+        isf_public_interface => build_isf_public_interface_contract(),
         serializable_plan_reports => build_serializable_plan_report_contract(),
         typed_extensions => build_extension_contract(),
         debug_runtime => build_debug_runtime_contract(),